+++ code/ft_memset.c
@@ -6,25 +6,25 @@
 /*   By: bbernhol <bbernhol@student.42wolfsburg.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/13 14:59:12 by bbernhol          #+#    #+#             */
-/*   Updated: 2022/05/13 15:16:49 by bbernhol         ###   ########.fr       */
+/*   Updated: 2022/06/06 17:08:23 by bbernhol         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <unistd.h>
 #include "libft.h"
 
-void *ft_memset(void *s, int c, unsigned long int n)
+void	*ft_memset(void *s, int c, unsigned long int n)
 {
-    char *cast_s;
-    int i;
+	char					*cast_s;
+	unsigned long int		i;
 
-    i = 1;
-    cast_s = (char*)s;
-    while (i <= n)
-    {
-        *cast_s = (char)c;
-        cast_s++;
-        i++;
-    }
-    return (s);
+	i = 1;
+	cast_s = (char *)s;
+	while (i <= n)
+	{
+		*cast_s = (char)c;
+		cast_s++;
+		i++;
+	}
+	return (s);
 }
